pythonProjectStreamlit/requirements.txt
joblib~=1.5.1
azure-storage-blob
pandas~=2.2.3
streamlit~=1.46.0
scikit-learn
xgboost~=2.1.4